uapi: convert utsname.h
