+++ include/protura/utsname.h
@@ -8,21 +8,9 @@
 #ifndef INCLUDE_PROTURA_UTSNAME_H
 #define INCLUDE_PROTURA_UTSNAME_H
 
-#include <protura/types.h>
-
-struct utsname {
-    char sysname[65];
-    char nodename[65];
-    char release[65];
-    char version[65];
-    char machine[65];
-};
-
-#ifdef __KERNEL__
+#include <uapi/protura/utsname.h>
 
 int sys_uname(struct user_buffer utsname);
 extern struct procfs_entry_ops proc_version_ops;
-
-#endif
 
 #endif
