+++ include/uapi/protura/utsname.h
@@ -1,0 +1,21 @@
+/*
+ * Copyright (C) 2020 Matt Kilgore
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License v2 as published by the
+ * Free Software Foundation.
+ */
+#ifndef __INCLUDE_UAPI_PROTURA_UTSNAME_H__
+#define __INCLUDE_UAPI_PROTURA_UTSNAME_H__
+
+#include <protura/types.h>
+
+struct utsname {
+    char sysname[65];
+    char nodename[65];
+    char release[65];
+    char version[65];
+    char machine[65];
+};
+
+#endif
